Initial commit for blipp aggregation.
More testing is needed, but everything seems to be working.
The current implementation of blipp is not using enough timeseries
events, and the calipers implementation needs to adjust the schema
to the one blipp is using. We can then refactor this into a general
purpose timeseries based aggregation transformation step.



git-svn-id: https://damsl.cs.indiana.edu/svn/xsp/trunk@189 6dc24a1a-a86e-4cf8-a923-33330a3f456b