working on MacOSX 10.7
Changes to config and compat to work on MacOSX.

There were some changes in the linking in that things that can be delayed
for runtime linking really need to find their symbols at library link time.

module_info() is commented out in xsp_auth_pass_{file,sqlite} as it
conflicted with the symbol in xsp_auth_pass, but it seems like those
two modules depend on things in the base auth_pass.